oauth/pyproject.toml
[project]
name = "oauth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.136.1",
    "passlib>=1.7.4",
    "pydantic>=2.13.3",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.26",
    "requests>=2.33.1",
    "uvicorn>=0.46.0",
]
